C:/Program Files/Git/generatePayoff for individual payoff

+++ src/com/dao/DerivativeDAO.java
@@ -14,6 +14,6 @@ public interface DerivativeDAO {
 	public boolean updateUserHolding(String emailId,String type,String symbol,String position,String expiryDate,double strikePrice,int numLots,double spotPrice,double premium);
 	public int deleteHolding(String emailId,String symbol,String type,String expiryDate,double strikePrice);
 	public List<Holding> getSpecificDerivative(String symbol);
-	public double getLotSize(String symbol);
+	public int getLotSize(String symbol);
 	public double getLCP(String symbol,String type,String expiryDate,double strikePrice);
 }
